[ref-01] separate more the responsibilities mostly with modal
