+++ package.json
@@ -10,7 +10,7 @@
     "@types/react": "^18.2.13",
     "@types/react-dom": "^18.2.6",
     "babel-plugin-syntax-dynamic-import": "^6.18.0",
-    "bigbluebutton-html-plugin-sdk": "0.0.93",
+    "bigbluebutton-html-plugin-sdk": "0.0.94",
     "path": "^0.12.7",
     "react-intl": "^6.6.8",
     "react": "^18.2.0",
