+++ public/locales/index.json
@@ -1,0 +1,7 @@
+[
+  "en.json",
+  "fr-FR.json",
+  "fr.json",
+  "index.json",
+  "pt-BR.json"
+]
