+++ src/main/hooks.ts
@@ -1,11 +1,9 @@
 import {
-  ActionButtonDropdownOption,
-  ActionButtonDropdownSeparator,
   DataChannelTypes,
   GenericContentSidekickArea,
   PluginApi,
 } from 'bigbluebutton-html-plugin-sdk';
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import {
   createIntl, createIntlCache, defineMessages, IntlShape,
 } from 'react-intl';
@@ -20,14 +18,6 @@ const LOCALE_REQUEST_OBJECT = (!process.env.NODE_ENV || process.env.NODE_ENV ===
   } : undefined;
 
 const intlMessages = defineMessages({
-  writeCC: {
-    id: 'plugin.actionButtonDropdown.write',
-    description: 'action button dropdown label to start writing',
-  },
-  stopCC: {
-    id: 'plugin.actionButtonDropdown.remove',
-    description: 'action button dropdown label to start writing',
-  },
   sectionName: {
     id: 'plugin.actionButtonDropdown.sidekickComponent.sectionName',
     description: 'name of the sidekick component section',
@@ -58,83 +48,6 @@ export const useGetInternationalization = (pluginApi: PluginApi) => {
   };
 };
 
-export const useActionsButtonManager = (
-  pluginApi: PluginApi,
-  intl: IntlShape | null,
-  localeMessagesLoading: boolean,
-) => {
-  const {
-    data: activeCaptionMenusResponseFromDataChannel,
-    deleteEntry: deleteActiveCaptionMenuResponseFromDataChannel,
-  } = pluginApi.useDataChannel!<ActiveCaptionMenuInformation>(
-    'typed-captions-data-channel',
-    DataChannelTypes.ALL_ITEMS,
-    'caption-menus',
-  );
-
-  const [captionLocale, setCaptionLocale] = useState('');
-
-  const currentUserResponse = pluginApi.useCurrentUser!();
-
-  const [isModalOpen, setIsModalOpen] = useState(false);
-
-  const onRequestClose = () => {
-    setIsModalOpen(false);
-  };
-  useEffect(() => {
-    if (currentUserResponse?.data?.role === 'MODERATOR') {
-      let captionLocaleFromMenus = '';
-      if (captionLocale === '') {
-        activeCaptionMenusResponseFromDataChannel?.data?.forEach((item) => {
-          if (item.fromUserId === currentUserResponse?.data?.userId) {
-            setCaptionLocale(item.payloadJson.captionLocale);
-            captionLocaleFromMenus = item.payloadJson.captionLocale;
-          }
-        });
-      }
-      const entryIdToRemove = activeCaptionMenusResponseFromDataChannel?.data?.filter(
-        (item) => item.payloadJson.captionLocale === captionLocale
-        || captionLocaleFromMenus === item.payloadJson.captionLocale,
-      )[0]?.entryId || '';
-      let actionButtonDropdownOnClick = () => {
-        deleteActiveCaptionMenuResponseFromDataChannel([entryIdToRemove]);
-      };
-      let actionButtonDropdownLabel = '';
-      if (intl) {
-        if (!entryIdToRemove) {
-          actionButtonDropdownLabel = intl.formatMessage(intlMessages.writeCC);
-          actionButtonDropdownOnClick = () => {
-            setIsModalOpen(true);
-          };
-        } else actionButtonDropdownLabel = intl.formatMessage(intlMessages.stopCC);
-      }
-      if (!localeMessagesLoading && intl) {
-        pluginApi.setActionButtonDropdownItems([
-          new ActionButtonDropdownSeparator(),
-          new ActionButtonDropdownOption({
-            icon: 'closed_caption',
-            label: actionButtonDropdownLabel,
-            tooltip: 'this is a button injected by plugin',
-            allowed: true,
-            onClick: actionButtonDropdownOnClick,
-          }),
-        ]);
-      }
-    } else {
-      pluginApi.setActionButtonDropdownItems([]);
-      pluginApi.setGenericContentItems([]);
-    }
-  }, [
-    currentUserResponse,
-    activeCaptionMenusResponseFromDataChannel,
-    localeMessagesLoading, intl]);
-
-  return {
-    isModalOpen,
-    onRequestClose,
-  };
-};
-
 export const useTypedCaptionsPanelManager = (
   pluginApi: PluginApi,
   intl: IntlShape | null,
@@ -147,7 +60,6 @@ export const useTypedCaptionsPanelManager = (
 
   const currentUserResponse = pluginApi.useCurrentUser!();
 
-  /// contentFunction, name, section, buttonIcon
   useEffect(() => {
     if ((intl && !localeMessagesLoading)
       && activeCaptionMenusResponseFromDataChannel?.data
@@ -162,7 +74,7 @@ export const useTypedCaptionsPanelManager = (
           }),
           buttonIcon: 'closed_caption',
           section: sectionName,
-          open: true,
+          open: false,
           contentFunction: renderComponent(
             pluginUuid,
             intl,
